+++ ui/src/app/cohort-search/modifier-page/modifier-page.component.spec.tsx
@@ -2,7 +2,7 @@ import {mount, shallow} from 'enzyme';
 
 import {registerApiClient} from 'app/services/swagger-fetch-clients';
 import {serverConfigStore} from 'app/utils/stores'
-import {currentCohortSearchContextStore, currentWorkspaceStore} from 'app/utils/navigation';
+import {currentCohortSearchContextStore, currentWorkspaceStore, urlParamsStore} from 'app/utils/navigation';
 import {CohortBuilderApi, Domain, ModifierType, WorkspacesApi} from 'generated/fetch';
 import * as React from 'react';
 import {waitOneTickAndUpdate} from 'testing/react-test-helpers';
@@ -10,7 +10,6 @@ import {CohortBuilderServiceStub} from 'testing/stubs/cohort-builder-service-stu
 import {WorkspaceStubVariables} from 'testing/stubs/workspaces';
 import {workspaceDataStub} from 'testing/stubs/workspaces';
 import {WorkspacesApiStub} from 'testing/stubs/workspaces-api-stub';
-import {urlParamsStore} from '../../utils/url-params-store';
 import {ModifierPage} from './modifier-page.component';
