+++ ui/src/app/cohort-search/overview/overview.component.tsx
@@ -17,7 +17,7 @@ import colors, {colorWithWhiteness} from 'app/styles/colors';
 import {reactStyles, withCdrVersions, withCurrentWorkspace} from 'app/utils';
 import {triggerEvent} from 'app/utils/analytics';
 import {isAbortError} from 'app/utils/errors';
-import {currentWorkspaceStore, navigate, navigateByUrl} from 'app/utils/navigation';
+import {currentWorkspaceStore, navigate, navigateByUrl, urlParamsStore} from 'app/utils/navigation';
 import {WorkspaceData} from 'app/utils/workspace-data';
 import {
   AgeType,
@@ -28,7 +28,6 @@ import {
   TemporalTime,
   WorkspaceAccessLevel
 } from 'generated/fetch';
-import {urlParamsStore} from '../../utils/url-params-store';
 
 const COHORT_TYPE = 'AoU_Discover';
