+++ ui/src/app/components/app-router.tsx
@@ -1,9 +1,8 @@
-import {navigate, routeConfigDataStore} from 'app/utils/navigation';
+import {navigate, routeConfigDataStore, urlParamsStore} from 'app/utils/navigation';
 import {routeDataStore} from 'app/utils/stores';
 import * as fp from 'lodash/fp';
 import * as React from 'react';
 import { BrowserRouter, Link, Redirect, Route, Switch, useHistory, useLocation, useParams, useRouteMatch} from 'react-router-dom';
-import {urlParamsStore} from '../utils/url-params-store';
 
 const {Fragment} = React;
