+++ ui/src/app/components/breadcrumb.spec.tsx
@@ -1,5 +1,5 @@
 import {getTrail} from 'app/components/breadcrumb'
-import {BreadcrumbType, currentWorkspaceStore} from 'app/utils/navigation';
+import {BreadcrumbType, currentWorkspaceStore, urlParamsStore} from 'app/utils/navigation';
 import {registerApiClient} from "app/services/swagger-fetch-clients";
 
 import {WorkspacesApi} from "generated/fetch";
@@ -8,7 +8,6 @@ import {exampleCohortStubs} from "testing/stubs/cohorts-api-stub";
 import {ConceptSetsApiStub} from "testing/stubs/concept-sets-api-stub";
 import {workspaceDataStub, WorkspaceStubVariables} from 'testing/stubs/workspaces';
 import {WorkspacesApiStub} from 'testing/stubs/workspaces-api-stub';
-import {urlParamsStore} from '../utils/url-params-store';
 
 describe('getTrail', () => {
   beforeEach(() => {
