+++ ui/src/app/pages/data/cohort-review/detail-tabs.component.tsx
@@ -14,7 +14,7 @@ import {Domain, FilterColumns} from 'generated/fetch';
 import {TabPanel, TabView} from 'primereact/tabview';
 import {Observable} from 'rxjs/Observable';
 import {from} from 'rxjs/observable/from';
-import {urlParamsStore} from '../../../utils/url-params-store';
+import {urlParamsStore} from "../../../utils/navigation";
 
 const styles = reactStyles({
   container: {
