+++ ui/src/app/pages/data/cohort-review/table-page.tsx
@@ -21,7 +21,7 @@ import colors from 'app/styles/colors';
 import {datatableStyles} from 'app/styles/datatable';
 import {reactStyles, withCurrentWorkspace} from 'app/utils';
 import {triggerEvent} from 'app/utils/analytics';
-import {navigate, navigateByUrl} from 'app/utils/navigation';
+import {navigate, navigateByUrl, urlParamsStore} from 'app/utils/navigation';
 import {WorkspaceData} from 'app/utils/workspace-data';
 import {
   CohortStatus,
@@ -32,7 +32,6 @@ import {
   ParticipantCohortStatus,
   SortOrder,
 } from 'generated/fetch';
-import {urlParamsStore} from '../../../utils/url-params-store';
 
 const fields = [
   {field: 'participantId', name: 'Participant ID'},
