+++ ui/src/app/utils/navigation.tsx
@@ -23,6 +23,7 @@ export const currentWorkspaceStore = new BehaviorSubject<WorkspaceData>(undefine
 export const currentCohortStore = new BehaviorSubject<Cohort>(undefined);
 export const currentConceptSetStore = new BehaviorSubject<ConceptSet>(undefined);
 export const globalErrorStore = new BehaviorSubject<ErrorResponse>(undefined);
+export const urlParamsStore = new BehaviorSubject<any>({});
 export const queryParamsStore = new BehaviorSubject<any>({});
 export const routeConfigDataStore = new BehaviorSubject<any>({});
 export const currentCohortCriteriaStore = new BehaviorSubject<Array<Selection>>(undefined);
