+++ ui/src/test.ts
@@ -13,9 +13,15 @@ import {
   platformBrowserDynamicTesting
 } from '@angular/platform-browser-dynamic/testing';
 import {cohortReviewStore} from 'app/services/review-state.service';
-import {currentWorkspaceStore, currentCohortStore, currentConceptSetStore, queryParamsStore, routeConfigDataStore} from 'app/utils/navigation';
+import {
+  currentWorkspaceStore,
+  currentCohortStore,
+  currentConceptSetStore,
+  queryParamsStore,
+  routeConfigDataStore,
+  urlParamsStore
+} from 'app/utils/navigation';
 import {serverConfigStore} from "./app/utils/stores";
-import {urlParamsStore} from './app/utils/url-params-store';
 
 // Unfortunately there's no typing for the `__karma__` variable. Just declare it as any.
 declare let __karma__: any;
